+++ zpa/resource_zpa_policy_access_inspection_rule.go
@@ -159,7 +159,6 @@ func resourcePolicyInspectionRuleUpdate(d *schema.ResourceData, meta interface{}
 	if err != nil {
 		return err
 	}
-	// Replace ValidatePolicyRuleConditions with ValidateConditions
 	if err := ValidateConditions(req.Conditions, zClient, GetString(d.Get("microtenant_id"))); err != nil {
 		return err
 	}