+++ zpa/resource_zpa_policy_access_rule_v2.go
@@ -63,7 +63,7 @@ func resourcePolicyAccessRuleV2() *schema.Resource {
 				Description: "This is for providing a customer message for the user.",
 			},
 			"conditions": {
-				Type:        schema.TypeList,
+				Type:        schema.TypeSet,
 				Optional:    true,
 				Computed:    true,
 				Description: "This is for proviidng the set of conditions for the policy.",
@@ -83,7 +83,7 @@ func resourcePolicyAccessRuleV2() *schema.Resource {
 							}, false),
 						},
 						"operands": {
-							Type:        schema.TypeList,
+							Type:        schema.TypeSet,
 							Optional:    true,
 							Computed:    true,
 							Description: "This signifies the various policy criteria.",
@@ -122,7 +122,7 @@ func resourcePolicyAccessRuleV2() *schema.Resource {
 										}, false),
 									},
 									"entry_values": {
-										Type:     schema.TypeList,
+										Type:     schema.TypeSet,
 										Optional: true,
 										Computed: true,
 										Elem: &schema.Resource{