+++ docs/guides/release-notes.md
@@ -12,9 +12,19 @@ Track all ZPA Terraform provider's releases. New resources, features, and bug fi
 
 ---
 
-``Last updated: v3.32.5``
+``Last updated: v3.32.6``
 
 ---
+
+## 3.32.6 (August, xx 2024)
+
+### Notes
+
+- Release date: **(August, xx 2024)**
+- Supported Terraform version: **v1.x**
+
+### Bug Fixes
+- [PR #482](https://github.com/zscaler/terraform-provider-zpa/pull/482) - Fixed drift issue within all policy access rule v2 resources.
 
 ## 3.32.5 (August, 28 2024)